[refactoring] application/shared/net/http: rename requests/connecting to requests/connections. update dependent modules.
