+++ src/app/application/shared/net/http/requests/connections/Types.js
