+++ src/app/application/utils/withEntity.js
@@ -4,7 +4,7 @@ import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { createRecordGetter } from '../shared/net/http/records/Repository';
 import { hasQueryMetaResult } from '../shared/net/http/requests/queries/Repository';
-import { createConnectionChecker } from '../shared/net/http/requests/connecting/Repository';
+import { createConnectionChecker } from '../shared/net/http/requests/connections/Repository';
 import { createErrorGetter } from '../shared/net/http/requests/errors/Repository';
 
 export default ({ readResource, requestId, resourceType }) => (
